Revert change requiring lastProcessedHeight on store cache

+++ packages/node/src/configure/SchemaMigration.service.test.ts
@@ -40,7 +40,9 @@ describe('SchemaMigration integration tests', () => {
   });
 
   afterAll(async () => {
-    await rimraf(tempDir);
+    if (tempDir) {
+      await rimraf(tempDir);
+    }
     await sequelize?.close();
   });
